Edited services and branch offices controllers and added approval for accounts.

+++ VerzijaBezAppUser/RentApp/RentApp/Helpers/ImageHelper.cs
@@ -82,7 +82,10 @@ namespace RentApp.Helpers
         public static void DeleteImage(string imageId)
         {
             string filePath = HostingEnvironment.MapPath("~/App_Data/" + imageId);
-            File.Delete(filePath);
+            if (File.Exists(filePath))
+            {
+                File.Delete(filePath);
+            }
         }
     }
 }
